+++ lib/graphql_rails/types/hidable_by_group.rb
@@ -6,10 +6,11 @@ module GraphqlRails
   module Types
     # Add visibility option based on groups
     module HidableByGroup
-      def initialize(*args, groups: [], **kwargs, &block)
+      def initialize(*args, groups: [], hidden_in_groups: [], **kwargs, &block)
         super(*args, **kwargs, &block)
 
-        @groups = groups.map(&:to_s)
+        @hidden_in_groups = hidden_in_groups.map(&:to_s)
+        @groups = groups.map(&:to_s) - @hidden_in_groups
       end
 
       def visible?(context)
@@ -22,10 +23,29 @@ module GraphqlRails
         @groups
       end
 
+      def hidden_in_groups
+        @hidden_in_groups
+      end
+
       def visible_in_context_group?(context)
+        group = context_graphql_group(context)
+
+        return true if no_visibility_configuration?(group)
+        return groups.include?(group) unless groups.empty?
+
+        !hidden_in_groups.include?(group)
+      end
+
+      def no_visibility_configuration?(group)
+        return true if group.nil?
+
+        groups.empty? && hidden_in_groups.empty?
+      end
+
+      def context_graphql_group(context)
         group = context[:graphql_group] || context['graphql_group']
 
-        group.nil? || groups.empty? || groups.include?(group.to_s)
+        group.nil? ? nil : group.to_s
       end
     end
   end